+++ lib/crosspost.dart
@@ -1,9 +1,9 @@
 library crosspost;
 
+export 'package:crosspost/src/common/exception.dart';
 export 'package:crosspost/src/common/social_content.dart' show SocialContent;
 export 'package:crosspost/src/common/social_content_data.dart'
     show TextContentData;
-//export 'package:crosspost/src/common/social_gateway.dart';
 export 'package:crosspost/src/common/social_post.dart' show SocialPost;
 export 'package:crosspost/src/gateway/telegram/telegram_gateway.dart'
     show TelegramGateway, TelegramResponse;
