+++ lib/develop.dart
@@ -1,3 +1,7 @@
 /// Develop library for tests and extending functionality.
 
 library crosspose.develop;
+
+export 'package:crosspost/crosspost.dart';
+export 'package:crosspost/src/common/social_content.dart' show ISocialContent;
+export 'package:crosspost/src/common/social_gateway.dart' show ISocialGateway;
